requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via uvicorn
asyncio-nats-client==0.11.4
    # via -r requirements.in
click==8.0.1
    # via uvicorn
fastapi==0.68.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
pydantic==1.8.2
    # via fastapi
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.2
    # via pydantic
uvicorn==0.15.0
    # via -r requirements.in
